path handling and cleanup
