+++ webui/components/chat/input/progress.html
