+++ webui/components/modals/full-screen-input.html
@@ -1,7 +1,7 @@
 <html>
 <head>
   <script type="module">
-    import { store } from "/components/chat/input/full-screen-store.js";
+    import { store } from "/components/modals/full-screen-store.js";
   </script>
 </head>
 <body>
