+++ webui/components/sidebar/chats/chats-list.html
@@ -1,7 +1,7 @@
 <html>
 <head>
   <script type="module">
-    import { store } from "/components/sidebar/chats-store.js";
+    import { store } from "/components/sidebar/chats/chats-store.js";
   </script>
 </head>
 <body>
