+++ webui/components/sidebar/left-sidebar.html
@@ -1,48 +1,82 @@
 <html>
 <head>
   <script type="module">
-    import { store } from "/components/app/app-store.js";
+    import { store as sidebarStore } from "/components/sidebar/sidebar-store.js";
   </script>
   <!-- Wrapper composes existing sidebar components -->
 </head>
 <body>
-  <!--Sidebar upper elements-->
-  <div class="left-panel-top">
-    <x-component path="sidebar/quick-actions.html"></x-component>
+  <div id="left-panel" class="panel" x-data :class="{'hidden': !$store.sidebar.isOpen}">
+    <!--Sidebar upper elements-->
+    <div class="left-panel-top">
+      <x-component path="sidebar/top-section/quick-actions.html"></x-component>
 
-    <!-- Tabs container -->
-    <x-component path="sidebar/tabs.html"></x-component>
+      <!-- Tabs container -->
+      <x-component path="sidebar/tabs.html"></x-component>
 
-    <!-- Chats List -->
-    <div class="config-section" id="chats-section" x-data x-show="$store.chats?.contexts?.length > 0 || true">
-      <x-component path="sidebar/chats-list.html"></x-component>
+      <!-- Chats List -->
+      <div class="config-section" id="chats-section" x-data x-show="$store.chats?.contexts?.length > 0 || true">
+        <x-component path="sidebar/chats/chats-list.html"></x-component>
+      </div>
+
+      <!-- Tasks List -->
+      <div class="config-section" id="tasks-section" x-data style="display: none;">
+        <x-component path="sidebar/tasks/tasks-list.html"></x-component>
+      </div>
     </div>
-
-    <!-- Tasks List -->
-    <div class="config-section" id="tasks-section" x-data style="display: none;">
-      <x-component path="sidebar/tasks-list.html"></x-component>
-    </div>
-  </div>
-  <!--Sidebar lower elements-->
-  <div class="left-panel-bottom">
-    <!-- Preferences -->
-    <x-component path="sidebar/preferences/preferences-panel.html"></x-component>
-    <!-- Version Info -->
-    <div class="version-info" x-data>
-      <template x-if="$store.app">
-        <span id="a0version" x-text="$store.app.versionLabel"></span>
-      </template>
+    <!--Sidebar lower elements-->
+    <div class="left-panel-bottom">
+      <!-- Preferences -->
+      <x-component path="sidebar/preferences/preferences-panel.html"></x-component>
+      <!-- Version Info -->
+      <div class="version-info" x-data>
+        <template x-if="$store.sidebar">
+          <span id="a0version" x-text="$store.sidebar.versionLabel"></span>
+        </template>
+      </div>
     </div>
   </div>
 
   <script>
     document.addEventListener('alpine:init', () => {
-      const s = Alpine.store('app');
-      if (s && typeof s.init === 'function') s.init();
+        const sidebarStore = Alpine.store('sidebar');
+        if (sidebarStore && typeof sidebarStore.init === 'function') sidebarStore.init();
     });
   </script>
 
   <style>
+    /* Left Panel styles from index.css */
+    #left-panel {
+      background-color: var(--color-panel);
+      border-right: 1px solid var(--color-border);
+      box-sizing: border-box;
+      display: flex;
+      flex-direction: column;
+      justify-content: space-between;
+      transition: all var(--transition-speed) ease-in-out;
+      width: 250px;
+      min-width: 250px;
+      color: var(--color-text);
+      box-shadow: 1px 0 5px rgba(0, 0, 0, 0.3);
+      user-select: none;
+    }
+
+    /* Ensure component host behaves like the original direct children for flex layout */
+    #left-panel > .left-panel-top,
+    #left-panel > .left-panel-bottom {
+        display: flex;
+        flex-direction: column;
+    }
+
+    #left-panel > .left-panel-top {
+        flex: 1;
+        min-height: 0; /* allow inner flex children to shrink properly */
+    }
+
+    #left-panel.hidden {
+      margin-left: -250px;
+    }
+
     .left-panel-top {
       flex: 1;
       display: -webkit-flex;
@@ -80,6 +114,23 @@
     /* Version label area */
     .version-info { line-height: 0.8rem; position: relative; margin: 0 var(--spacing-md) 1rem var(--spacing-md); padding-top: 10px; border-top: 1px solid var(--color-border); }
     #a0version { color: var(--color-text); opacity: 0.7; font-size: 0.7rem; user-select: all; }
+    @media (max-width: 768px) {
+      #left-panel {
+        position: fixed;
+        left: 0;
+        top: 0;
+        bottom: 0;
+        width: 250px !important; /* Force width */
+        min-width: 250px;
+        z-index: 1003;
+        transition: all var(--transition-speed) ease-in-out;
+      }
+
+      #left-panel.hidden {
+        margin-left: -250px;
+      }
+    }
+
     @media (max-height: 600px) {
       #chats-section { min-height: 100%; }
       .left-panel-top { overflow-y: auto; -webkit-scroll-behavior: smooth; scroll-behavior: smooth; }
