+++ webui/components/sidebar/tasks/tasks-store.js
