+++ webui/index.css
@@ -115,7 +115,6 @@ p:last-child {
 }
 
 body,
-#left-panel,
 #chat-input,
 .message,
 .config-button,
@@ -149,41 +148,6 @@ img {
   overflow: auto;
   -webkit-scroll-behavior: smooth;
   scroll-behavior: smooth;
-}
-
-/* Left Panel */
-#left-panel {
-  background-color: var(--color-panel);
-  border-right: 1px solid var(--color-border);
-  box-sizing: border-box;
-  display: -webkit-flex;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  -webkit-transition: all var(--transition-speed) ease-in-out;
-  transition: all var(--transition-speed) ease-in-out;
-  width: 250px;
-  min-width: 250px;
-  color: var(--color-text);
-  box-shadow: 1px 0 5px rgba(0, 0, 0, 0.3);
-  user-select: none;
-}
-
-/* Ensure component host behaves like the original direct children for flex layout */
-#left-panel > x-component {
-  display: -webkit-flex;
-  display: flex;
-  flex-direction: column;
-  flex: 1;
-  min-height: 0; /* allow inner flex children to shrink properly */
-}
-
-#left-panel.hidden {
-  margin-left: -250px;
-}
-
-.switch-label {
-  margin-right: 0.5rem;
 }
 
 /* Right Panel */
@@ -920,22 +884,6 @@ input:checked + .slider:before {
 }
 
 @media (max-width: 768px) {
-  #left-panel {
-    position: fixed;
-    left: 0;
-    top: 0;
-    bottom: 0;
-    width: 250px !important; /* Force width */
-    min-width: 250px;
-    z-index: 1003;
-    -webkit-transition: all var(--transition-speed) ease-in-out;
-    transition: all var(--transition-speed) ease-in-out;
-  }
-
-  #left-panel.hidden {
-    margin-left: -250px;
-  }
-
   /* .copy-button {
     display: none !important;
   } */
@@ -1023,10 +971,6 @@ a:active {
 
 .light-mode .disconnected {
   color: #f44336;
-}
-
-.light-mode #left-panel {
-  box-shadow: 1px 0 25px rgba(0, 0, 0, 0.05);
 }
 
 .light-mode .config-button {
